+++ docs/source/whatsnew.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.5]
+
+### Changed
+- ``twoaxistracking.__version__`` now correctly reports the version string instead
+  of raising ``AttributeError``. (:pull:`45`)
+
+
 ## [0.2.4] - 2023-01-05
 
 ### Changed